rename vumeter as ts and tsx file were conflicting on OSX (only differ in casing)
